style(rename): db.images -> db.pictures
