+++ backend/matcha/services/pictures.py
@@ -7,7 +7,7 @@ from werkzeug.utils import secure_filename
 
 from matcha.app_utils import check_request_json, make_unique
 
-from matcha.db.images import (
+from matcha.db.pictures import (
     db_get_user_images,
     db_upload_pictures,
     db_set_profile_picture,
