+++ backend/matcha/services/users.py
@@ -9,7 +9,7 @@ from flask_jwt_extended import (
 
 from matcha.app_utils import check_request_json
 
-from matcha.db.images import (
+from matcha.db.pictures import (
     db_get_user_images,
     db_get_url_profile,
 )
